+++ R/module-filterDF.R
@@ -93,7 +93,7 @@ filterDF <- function(input, output, session,
     filter_inputs <- lapply(
       X = rv_filters$mapping, 
       FUN = function(x) {
-        req(input[[x]])
+        # req(input[[x]])
         input[[x]]
       }
     )
@@ -150,16 +150,22 @@ create_filters <- function(data, vars = NULL, width = "100%", session = getDefau
     X = vars,
     FUN = function(variable) {
       var <- data[[variable]]
+      any_na <- anyNA(var)
       var <- var[!is.na(var)]
       id <- filters_id[[variable]]
+      tag_label <- if (any_na) {
+        tags$span(
+          tags$label(variable), HTML("&nbsp;&nbsp;"), 
+          na_filter(id = ns(paste0("na_", id)))
+        )
+      } else {
+        tags$span(tags$label(variable), HTML("&nbsp;&nbsp;"))
+      }
       if (inherits(x = var, what = c("numeric", "integer"))) {
         params <- find_range_step(var)
         tags$div(
           style = "position: relative;",
-          tags$span(
-            tags$label(variable), HTML("&nbsp;&nbsp;"), 
-            na_filter(id = ns(paste0("na_", id)))
-          ),
+          tag_label,
           set_slider_attr(sliderInput(
             inputId = ns(id), 
             min = params$min, 
@@ -174,10 +180,7 @@ create_filters <- function(data, vars = NULL, width = "100%", session = getDefau
         range_var <- range(var)
         tags$div(
           style = "position: relative;",
-          tags$span(
-            tags$label(variable), HTML("&nbsp;&nbsp;"), 
-            na_filter(id = ns(paste0("na_", id)))
-          ),
+          tag_label,
           set_slider_attr(sliderInput(
             inputId = ns(id), 
             min = min(var), 
@@ -193,10 +196,7 @@ create_filters <- function(data, vars = NULL, width = "100%", session = getDefau
         tags$div(
           style = "position: relative;",
           class = if (length(values) > 15) "selectize-big",
-          tags$span(
-            tags$label(variable), HTML("&nbsp;&nbsp;"), 
-            na_filter(id = ns(paste0("na_", id))) 
-          ),
+          tag_label,
           selectizeInput(
             inputId = ns(id),
             choices = values, 
@@ -257,48 +257,84 @@ make_expr_filter <- function(filters, filters_na, data, data_name) {
       values <- filters[[var]]
       nas <- filters_na[[var]]
       data_values <- data[[var]]
-      if (!match_class(values, data_values))
+      if (!is.null(values) & !match_class(values, data_values))
         return(NULL)
       values_expr <- NULL
       if (inherits(x = values, what = c("numeric", "integer"))) {
-        data_values <- find_range_step(data_values)$range
-        if (!isTRUE(all.equal(values, data_values))) {
+        data_range <- find_range_step(data_values)$range
+        if (!isTRUE(all.equal(values, data_range))) {
           if (isTRUE(nas)) {
-            values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2] | is.na(!!sym(var)))
+            if (anyNA(data_values)) {
+              values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2] | is.na(!!sym(var)))
+            } else {
+              values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2])
+            }
           } else {
-            values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2] & !is.na(!!sym(var)))
+            if (anyNA(data_values)) {
+              values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2] & !is.na(!!sym(var)))
+            } else {
+              values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2])
+            }
           }
         }
       } else if (inherits(x = values, what = c("Date", "POSIXct"))) {
         values <- format(values)
-        data_values <- range(data_values, na.rm = TRUE)
-        data_values <- format(data_values)
-        if (!identical(values, data_values)) {
+        data_range <- range(data_values, na.rm = TRUE)
+        data_range <- format(data_range)
+        if (!identical(values, data_range)) {
           if (isTRUE(nas)) {
-            values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2] | is.na(!!sym(var)))
+            if (anyNA(data_values)) {
+              values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2] | is.na(!!sym(var)))
+            } else {
+              values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2])
+            }
           } else {
-            values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2] & !is.na(!!sym(var)))
+            if (anyNA(data_values)) {
+              values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2] & !is.na(!!sym(var)))
+            } else {
+              values_expr <- expr(!!sym(var) >= !!values[1] & !!sym(var) <= !!values[2])
+            }
           }
         }
       } else {
         data_values <- unique(as.character(data_values))
         if (!identical(sort(values), sort(data_values))) {
-          if (length(values) <= length(data_values)/2) {
+          if (length(values) == 0) {
             if (isTRUE(nas)) {
-              values_expr <- expr(!!sym(var) %in% !!values | is.na(!!sym(var)))
+              values_expr <- expr(is.na(!!sym(var)))
             } else {
-              values_expr <- expr(!!sym(var) %in% !!values)
+              values_expr <- expr(!(!!sym(var) %in% !!data_values[!is.na(data_values)]) & !is.na(!!sym(var)))
             }
           } else {
-            if (isTRUE(nas)) {
-              values_expr <- expr(!(!!sym(var) %in% !!setdiff(data_values, values)) | is.na(!!sym(var)))
+            if (length(values) <= length(data_values)/2) {
+              if (isTRUE(nas)) {
+                if (anyNA(data_values)) {
+                  values_expr <- expr(!!sym(var) %in% !!values | is.na(!!sym(var)))
+                } else {
+                  values_expr <- expr(!!sym(var) %in% !!values)
+                }
+              } else {
+                values_expr <- expr(!!sym(var) %in% !!values)
+              }
             } else {
-              values_expr <- expr(!(!!sym(var) %in% !!setdiff(data_values, values)))
+              if (isTRUE(nas)) {
+                if (anyNA(data_values)) {
+                  values_expr <- expr(!(!!sym(var) %in% !!setdiff(data_values[!is.na(data_values)], values)) | is.na(!!sym(var)))
+                } else {
+                  values_expr <- expr(!(!!sym(var) %in% !!setdiff(data_values[!is.na(data_values)], values)))
+                }
+              } else {
+                if (anyNA(data_values)) {
+                  values_expr <- expr(!(!!sym(var) %in% !!setdiff(data_values[!is.na(data_values)], values)) & !is.na(!!sym(var)))
+                } else {
+                  values_expr <- expr(!(!!sym(var) %in% !!setdiff(data_values[!is.na(data_values)], values)))
+                }
+              }
             }
           }
         }
       }
-      if (is.null(values_expr) & !isTRUE(nas)) {
+      if (is.null(values_expr) & !isTRUE(nas) & anyNA(data_values)) {
         expr(!is.na(!!sym(var)))
       } else {
         values_expr